Rename FeatureExtraction to ExtractNuclearFeatures elsewhere
